feature/IN-489: Update Passport SAML package to handle customised Namespace qualifiers.
Fix to make the validate Response function independent of the Namesapce qualifiers. The fix involves using a regex to extract the hash keys in the SAMsapce qualifiers. The fix involves using a regex to extract the hash keys in the SAML response. This makes it independent of the customized namespace qualifiers.