bugfix/KB-5923: Fix to handle different saml versions
Due to different saml versions which we may receive from a saml response provider (IDPs), we need to support keys which have prefix saml or saml2.
This was being done using a regexp. But the regexp match was flawed and failed for some scenarios.
This has been fixed using a simpler approach.